Renamed RuntimeUpdates sample project for consistence
- both directory and project are now named SvgRuntimeUpdates
- fixes #773
