+++ Source/Svg.sln
@@ -19,7 +19,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SvgW3CTestRunner", "..\Test
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SVGViewer", "..\Samples\SVGViewer\SVGViewer.csproj", "{1B8F3C8A-CCAC-474E-B09D-522FBA93DCFD}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SvgSample", "..\Samples\RuntimeUpdates\SvgSample.csproj", "{9782B15C-D0BA-4A4E-9A03-F97082AF9256}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SvgRuntimeUpdates", "..\Samples\SvgRuntimeUpdates\SvgRuntimeUpdates.csproj", "{9782B15C-D0BA-4A4E-9A03-F97082AF9256}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Samples", "Samples", "{038E2E7B-BC03-4DA7-AA5F-CA8BD95BD0F2}"
 EndProject
@@ -71,6 +71,7 @@ Global
 		{8ED74C39-6CFF-421E-952A-30F9E6957108}.Release|Mixed Platforms.Build.0 = Release|x86
 		{8ED74C39-6CFF-421E-952A-30F9E6957108}.Release|x86.ActiveCfg = Release|x86
 		{8ED74C39-6CFF-421E-952A-30F9E6957108}.Release|x86.Build.0 = Release|x86
+		{8ED74C39-6CFF-421E-952A-30F9E6957108}.Debug|Any CPU.Build.0 = Debug|x86
 		{1B8F3C8A-CCAC-474E-B09D-522FBA93DCFD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{1B8F3C8A-CCAC-474E-B09D-522FBA93DCFD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1B8F3C8A-CCAC-474E-B09D-522FBA93DCFD}.Debug|Mixed Platforms.ActiveCfg = Debug|x86
@@ -93,6 +94,7 @@ Global
 		{9782B15C-D0BA-4A4E-9A03-F97082AF9256}.Release|Mixed Platforms.Build.0 = Release|x86
 		{9782B15C-D0BA-4A4E-9A03-F97082AF9256}.Release|x86.ActiveCfg = Release|x86
 		{9782B15C-D0BA-4A4E-9A03-F97082AF9256}.Release|x86.Build.0 = Release|x86
+		{9782B15C-D0BA-4A4E-9A03-F97082AF9256}.Debug|Any CPU.Build.0 = Debug|x86
 		{F34CF345-E7EF-4354-B49A-AFE7A78442A7}.Debug|Any CPU.ActiveCfg = Debug|x86
 		{F34CF345-E7EF-4354-B49A-AFE7A78442A7}.Debug|Mixed Platforms.ActiveCfg = Debug|x86
 		{F34CF345-E7EF-4354-B49A-AFE7A78442A7}.Debug|Mixed Platforms.Build.0 = Debug|x86
@@ -103,6 +105,7 @@ Global
 		{F34CF345-E7EF-4354-B49A-AFE7A78442A7}.Release|Mixed Platforms.Build.0 = Release|x86
 		{F34CF345-E7EF-4354-B49A-AFE7A78442A7}.Release|x86.ActiveCfg = Release|x86
 		{F34CF345-E7EF-4354-B49A-AFE7A78442A7}.Release|x86.Build.0 = Release|x86
+		{F34CF345-E7EF-4354-B49A-AFE7A78442A7}.Debug|Any CPU.Build.0 = Debug|x86
 		{8CEEB917-63E2-4AE7-B0D1-50A840050975}.Debug|Any CPU.ActiveCfg = Debug|x86
 		{8CEEB917-63E2-4AE7-B0D1-50A840050975}.Debug|Mixed Platforms.ActiveCfg = Debug|x86
 		{8CEEB917-63E2-4AE7-B0D1-50A840050975}.Debug|Mixed Platforms.Build.0 = Debug|x86
@@ -113,6 +116,7 @@ Global
 		{8CEEB917-63E2-4AE7-B0D1-50A840050975}.Release|Mixed Platforms.Build.0 = Release|x86
 		{8CEEB917-63E2-4AE7-B0D1-50A840050975}.Release|x86.ActiveCfg = Release|x86
 		{8CEEB917-63E2-4AE7-B0D1-50A840050975}.Release|x86.Build.0 = Release|x86
+		{8CEEB917-63E2-4AE7-B0D1-50A840050975}.Debug|Any CPU.Build.0 = Debug|x86
 		{9379360C-CB0B-4035-AE8F-8863A420FDCF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{9379360C-CB0B-4035-AE8F-8863A420FDCF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9379360C-CB0B-4035-AE8F-8863A420FDCF}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
